requirements.txt
torch~=2.6.0
pandas~=2.2.3
transformers~=4.49.0
pillow~=11.1.0
jellyfish~=1.2.0
requests~=2.32.3
pymupdf~=1.25.3
tqdm~=4.67.1
selenium~=4.29.0
matplotlib~=3.9.4
seaborn~=0.13.2
pydantic~=2.10.6
numpy~=1.26.4
scikit-learn~=1.6.1
uvicorn~=0.34.0
fastapi~=0.115.11
neo4j~=5.28.1
pymongo~=4.11.1
mistralai~=1.5.1
pymupdf4llm~=0.0.17
weaviate-client
sentence_transformers
webdriver_manager
openpyxl
google-genai
